Fix: Excluding old 'client_registration'
